Rollback and renamed SpotifyAuth
