+++ src/types.d.ts
@@ -1,0 +1,5 @@
+export interface AuthContextType{
+    accessToken: string | undefined;
+    refreshToken: string | undefined;
+    saveTokens: (accessToken: string | undefined, refreshToken: string | undefined) => void;
+}
